move forest, pinball to single
