+++ omni_drones/envs/single/__init__.py
@@ -25,4 +25,5 @@ from .hover import Hover
 from .track import Track
 from .trackV1 import TrackV1
 from .fly_through import FlyThrough
-from .nav import SDFNav
+from .forest import Forest
+from .pinball import Pinball
